Changes in KafkaTask
